LibFewShot/requirements.txt
einops>=0.3.0
future
matplotlib>=3.3.4
numpy>=1.19.2
pandas>=1.1.3
Pillow>=8.1.0
PyYAML>=5.4.1
rich>=9.3.0
scikit-learn>=0.24.1
scipy>=1.5.2
tensorboard>=2.4.1
optuna